Move Zotero functions to a separate file
